chore: Seperable Convolution Performance Exercise
The exercise is about analyzing performance of a seperable convolution layer over the alternate standard or normal convolution layer

A separable convolution is a convolutional layer that divides the standard convolutional layer into two layers: a depthwise convolution and a pointwise convolution. By doing so, the number of FLOPs or operations required to run the model is reduced.